+++ src/lib/model/position.ts
@@ -39,7 +39,7 @@ export default class Position extends Model<Position> {
   side: string;
 
   @Column({
-    type: DataType.FLOAT(10),
+    type: DataType.DECIMAL(20, 10),
     comment: '价格'
   })
   price: number;
@@ -51,7 +51,7 @@ export default class Position extends Model<Position> {
   mocktime: string;
 
   @Column({
-    type: DataType.INTEGER(10),
+    type: DataType.DECIMAL(20, 10),
     comment: '股数'
   })
   quantity: number;
